create dhis2_datasets (#4)
* create dhis2_datasets

functions to discover what are the datasets inside a DHIS2 server and to pull data stored in a remote DHIS2 server

* scripts moved to R/

Moved scripts to correct folder

* optimization

optmized code and added documentation

* updating dhis2 functions

major update that now allows datasets to be fetched even when there are variables that do not allow for aggregation

* Update dhis2_import.R

made some reviews and added a solution to retrieve indicators (and not stop when there are no indicators)

* changed package name and documentation

This is a try to change the package name. Check and Build tests were done and went fine after changing the package name. Folder name was not changed. Thanks @xmnlab !!!

* Revert "changed package name and documentation"

This reverts commit 320c408422864ea01150c02dcd37831d5b338fde.

* remove .DS_Store

* run check again

Just ran another test and it seems that there was some errors in package importing procedures. I corrected that now.

* Update r.yml

removed testing for R 3.6.1

Co-authored-by: Flávio Codeço Coelho <fccoelho@gmail.com>